requirements.txt
numpy==1.24.3
opencv_contrib_python==4.7.0.72
opencv_python==4.7.0.72
opencv_python_headless==4.8.0.74
Pillow==9.5.0
Pillow==10.2.0
streamlit==1.32.2
ultralytics==8.1.23
